Add missing interface text for "developer" group.
I enabled the "developer" group on my private testing server after seeing it in "DefaultSettings.php" (I enabled it by copying the entry to "LocalSettings.php") and noticed that it was missing the default interface text ("group-developer", "group-developer-member", and "grouppage-developer"). I would suggest adding these entries so that the "developer" group will have interface text by default if it is enabled on a wiki.