Revert actor changes to populateLogSearch.php
In all the past patches, we missed the fact that this maintenance script
is run as an upgrade from MW pre-1.16 to MW 1.16. It's apparently not
intended to be run outside of the upgrade process.

Thus, let's revert the changes that were made to the script in I8d825eb0
and I4764c1c7. Instead, the script need only detect that the pre-1.16
schema it's looking for doesn't exist anymore and bail out.

Bug: T188327
Change-Id: I7e03b45ef6d43b9069f0399089ab06e136a89fbc