RecentChangesUpdateJob: Fix GROUP BY
When a column specified in GROUP BY is both a table column and a
SELECT-field, MySQL and PostgreSQL both interpret it as the table column
rather than the SELECT-field.

In PostgreSQL this raises an error, which is good since it lets us know
it needs fixing. But MySQL goes ahead and groups by the table field
which gives us the wrong result.

Bug: T230618
Change-Id: Id500556b2795b86849329eece3b651b08e29a7f7
(cherry picked from commit 247904106076b80636ba9f3833a8d1e18be256a7)