ApiSandbox: Load mediawiki.hlist RL module
The default l10n for the main module's description uses the hlist class,
so load it.

It might make more sense to have it as a module dependency of an
appropriate RL module like mediawiki.apipretty instead of requiring
every caller to load it manually, but it seems that's not really done
because RL is weird about handling CSS-only modules as dependencies.

Bug: T230782
Change-Id: I6493d7d789f3cebee737645528cddfb1d3acb1d7