Cache nonexistence of grammar transformations
Previously every call to Language::getGrammarTransformations() for a
language without grammar transformations (i.e., almost all) would look
for a transformations file using is_readable(). This is not going to be
especially expensive, but it's presumably still more expensive than
caching the result.

The change is not actually testable right now without resorting to
TestingAccessWrapper or somehow mocking the file access, but I threw in
some basic tests for the method anyway. If later we decide how best to
mock the file access, they can be updated to check that the cache is
working properly.

Change-Id: I76e0fb7dbb1602df4f353722aacbd5adfff3c053