maintenance: Remove cross-wiki purging from rebuildmessages.php
This script relied on numerous outdated or unofficial methods that
are no longer supported.

* Global variable $messageMemc holding a BagOStuff based on
  MessageCacheType. This logic is now in ServiceWiring
  for the MessageCache service.

* String "{$db}:messages", should have been using BagOStuff::makeKey()
  and may've fallen out of sync.

* Deleting keys outright instead of touching check keys as
  MessageCache does, since it's been using WANObjectCache for
  several years.

All of this should be done via MessageCache::clear(), but that
can't be easily constructed across wikis on a wiki farm. That
requires instantiating the script for other wikis separately.

Remove support for that, recommending instead that site admins
run it in a loop for each wiki as needed.

Realistically though, that should never be needed. The script has
virtually no known use case, except (as documented in maintenance/README)
"after changing a wiki's content language" - which applies to one
wiki, not all a once.

Change-Id: I9f45d65f5cef93d6e332baf26d955d3f7a7c19d2