tests: Reset $wgContLang consistently in PHPUnit tests (take 2)
This a second attempt at 09407afd74104bb3a674e, which was reverted due
to breaking Wikibase integration tests.

The original commit had two separate typos that made it not work at all.
Additionally, some things were apparently expecting $wgContLang to be an
actual Language and not a StubObject. To solve that, we just set it
directly and don't use a StubObject. Initialization should be cheap, and
it would probably have been unstubbed on every test anyway.

Bug: T231799
Change-Id: I6d68a22e53e17458d0a63c13bd97c19baff4d1c7