rcfilters: Use default styling for the overlay
The class .oo-ui-defaultOverlay is used by OOUI, and all the skins
already define reasonable font-size and z-index for it, so we can
remove that here.

Keep the existing class .mw-rcfilters-ui-overlay because apparently
`right: 0; left: 0;` is required (that's weird…) and because Minerva
also wants to override it (I221f6c2d637c7d2a843bfc4a2f7cd9877e2a5b12).

Depends-On: I221f6c2d637c7d2a843bfc4a2f7cd9877e2a5b12
Change-Id: I672fdcebbdc586765de2c10f9a9a21796db57cc5