SpecialRunJobs: optional output stats and status.
This adds a parameter to SpecialRunJobs that lets it output statistics
about the jobs it has run.

The 'reached' field can be used to detect when the queue is emopty,
which is essential to know for clients that want to flush the entire
job queue, to ensure that all effects of any actions they have taken
have been processed.

More specifically, this provides a way for an external testing framework
to run all jobs after an action, so it can observe and assert the
effects of that action.

Bug: T231822
Change-Id: Ibb38490afca71efeb67300b9665951c429c19a3c