filebackend: optimize 'create', 'store', and 'copy' in FSFileBackend
Avoid latency of more race-condition prone is_file() calls.
Make the create/store/copy operations stream the file to a temp file
on the same device as the destination and then rename the temp file
over the destination file on success.

Use fopen() with stream_copy_to_stream() to avoid weird NFS copy()
bug and to be able to check the size properly without worrying about
race conditions since fstat() will be for the same inode that fopen()
worked on, even if the file was renamed over in the middle.

Since all operations to the "real" files happen via rename()/unlink()
or some script using the same underlying FS methods, this means that
FSFileBackend has isolation and atomicity closer to SwiftFileBackend
for these files.

Refactor the FSFileOpHandle/doExecuteOpHandlesInternal() chmod logic
to be in one place. Replace variables named $path with clearer $fsPath.

Change-Id: Ie07ec930b5cd2bcfea7b5aca11f6a667c6928e56